requirements.txt
# Core dependencies -- mirrors setup.py install_requires
panel>=1.3.0
python-dotenv>=1.0.0
openai>=1.0.0
tenacity>=8.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0

# Scientific / ML stack
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7

# Visualization and notebooks
hvplot>=0.9.0
plotly>=5.0.0
jupyter
ipykernel

# Azure infrastructure (ACI lifecycle management for TEI containers)
azure-mgmt-containerinstance>=10.0
azure-identity>=1.15

# Azure Blob Storage (artifact storage backend)
azure-storage-blob>=12.0

# Local Docker TEI management
docker>=7.0
